agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.14"
dependencies = [
    "bedrock-agentcore>=1.1.2",
    "boto3>=1.42.17",
    "google-cloud-aiplatform>=1.132.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "strands-agents>=1.20.0",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.2.5",
    "black>=25.12.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
